requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cssselect2==0.9.0
ctranslate2==4.7.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.129.0
faster-whisper==1.2.1
filelock==3.24.2
flatbuffers==25.12.19
fonttools==4.61.1
fpdf2==2.8.5
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
markdown2==2.5.4
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
mutagen==1.47.0
numpy==2.4.2
onnxruntime==1.24.1
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
pyphen==0.17.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.22
pywhispercpp==1.4.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
setuptools==82.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websockets==16.0
yt-dlp==2026.2.4
zopfli==0.4.1
